+++ list-service/src/main/java/ru/lakeevda/listservice/controller/ListController.java
@@ -20,7 +20,7 @@ public class ListController {
     }
 
     @GetMapping("/phone/{phone}")
-    public ResponseEntity<List<ListDto>> getLists(@PathVariable Integer phone) {
+    public ResponseEntity<List<ListDto>> getLists(@PathVariable long phone) {
         List<ListDto> lists = listService.getListByUserPhone(phone);
         return ResponseEntity.ok().body(lists);
     }
@@ -32,14 +32,14 @@ public class ListController {
     }
 
     @PutMapping("/phone/{phone}/update")
-    public ResponseEntity<Void> updateList(@PathVariable Integer phone,
+    public ResponseEntity<Void> updateList(@PathVariable long phone,
                                            @RequestBody ListDto list) {
         listService.updateList(list, phone);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/phone/{phone}/delete/{id}")
-    public ResponseEntity<Void> deleteList(@PathVariable Integer phone,
+    public ResponseEntity<Void> deleteList(@PathVariable long phone,
                                            @PathVariable Long id) {
         listService.deleteList(id, phone);
         return ResponseEntity.ok().build();
@@ -47,14 +47,14 @@ public class ListController {
 
     @PutMapping("/{id}/user/add")
     public ResponseEntity<ListDto> addUserToList(@PathVariable Long id,
-                                                 @RequestParam Integer phone) {
+                                                 @RequestParam long phone) {
         ListDto list = listService.addUserToList(id, phone);
         return ResponseEntity.ok().body(list);
     }
 
     @PutMapping("/{id}/user/delete")
     public ResponseEntity<ListDto> deleteUserFromList(@PathVariable Long id,
-                                                      @RequestParam Integer phone) {
+                                                      @RequestParam long phone) {
         ListDto list = listService.deleteUserFromList(id, phone);
         return ResponseEntity.ok().body(list);
     }
