+++ list-service/src/main/java/ru/lakeevda/listservice/entity/User.java
@@ -23,5 +23,5 @@ public class User {
     private String username;
     @Basic
     @Column(name = "phone", nullable = false)
-    private int phone;
+    private long phone;
 }
