+++ user-service/src/main/java/ru/lakeevda/userservice/entity/User.java
@@ -22,7 +22,7 @@ public class User {
     private String username;
     @Basic
     @Column(name = "phone", nullable = false)
-    private Integer phone;
+    private long phone;
     @Basic
     @Column(name = "email", nullable = true, length = 100)
     private String email;
